pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "drop-sentinel"
version = "0.1.0"
description = "Release monitoring and purchase assistant for Damai tickets and Pop Mart collectibles"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "httpx[http2]>=0.27",
    "pydantic>=2.0",
    "typer>=0.12",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
notify = [
    "aiogram>=3.0",
]
watch = [
    "apscheduler>=3.10",
]
auto = [
    "Appium-Python-Client>=4.0",
    "selenium>=4.20",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
]

[project.scripts]
drop-sentinel = "drop_sentinel.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/drop_sentinel"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
